feat: handle specific exception in case no element is found
Set an exception handler for the NoSuchElementException exception. In this case, return a specific message saying that no element has been found with a specific id.
Set the CREATED status in case of a successful call to the POST method, and the NO_CONTENT status in case of the DELETE method.
Remove the @GeneratedValue for the moment, as it does not work.

References: #7