docs: let the swagger ui be available for the REST service implemented
Add the org.springdoc webmvc dependency in the build.gradle.kts file for importing the necessary dependency to let the swagger ui be available via http call.
Change the context path to /kp/crud in the application.properties file, so that all the REST endpoints share such path, which is going to be not specific anymore.

References: #9