feat: add functions findBy() inside the TennisPlayerRepository class
Add the filters for the GET /tennisplayers endpoint: name, surname, and ranking. Based on this, define the functions findBy() using the indicated filters and the properties of the JPA library.
Always return a list of players as it is possible to have multiple elements in the response even with filters.

References: #13