feat: define the name, surname, and ranking query parameters in the GET /tennisplayers endpoint
Remove the description that is useful for documenting the open api, as it makes the code too heavy.
Add the query parameters name, surname, and ranking in the GET /tennisplayers endpoint, passing them to the service.
In the TennisPlayerService class, according to the presence or absence of the query parameters, call a different function of the repository interface.

References: #13