feat: apply the query parameters filtering via the CriteriaBuilder class
Implement the JpaSpecificationExecutor interface to the TennisPlayerRepository interface: this allows to add specifications to the find() functions, that correspond to WHERE conditions in the final query.
Define the TennisPlayerFilter class to add filter functions for each of the query parameters, and then group them together via AND connectors, using the Specifications.allOf() function.
In the TennisPlayerService class, exploit the applyFilters() method in the TennisPlayerFilter class.

References: #17